requirements.txt
Flask==2.1.0
flask-oidc==1.4.0
oauth2client==4.1.3
PyJWT==1.7.1
requests==2.22.0
itsdangerous==2.0.1
